+++ AngularApp/myApp/src/app/StyleClass/styleClass-main.component.ts
@@ -8,7 +8,7 @@ import { StyleClassService } from '../services/styleClass.service';
 })
 
 export class StyleClassMainComponent implements OnInit {
-  title = 'Style Classes';
+  title = 'Clases de estilo';
   styleClasses: Array<StyleClass>;
 
   constructor(private _styleClassService: StyleClassService) {
@@ -26,9 +26,14 @@ export class StyleClassMainComponent implements OnInit {
       ((obtainedStyleClasses: Array<StyleClass>) => this.styleClasses = obtainedStyleClasses),
       ((error: any) => console.log(error))
     )
+    window.location.reload();
   }
 
   deleteStyleClass(id: string): void {
-    this._styleClassService.deleteStyleClass(id).subscribe();
+    if (window.confirm('¿Esta seguro que desea eliminar el estilo?')) {
+      this._styleClassService.deleteStyleClass(id).subscribe();
+      alert("Clase de estilo eliminada");
+      window.location.reload();
+    }
   }
 }
